Renaming folder video_vpdq -> vpdq
