+++ python-threatexchange/threatexchange/extensions/vpdq/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
 
 from threatexchange.extensions.manifest import ThreatExchangeExtensionManifest
-from threatexchange.extensions.video_vpdq.video_vpdq import VideoVPDQSignal
+from threatexchange.extensions.vpdq.video_vpdq import VideoVPDQSignal
 
 
 TX_MANIFEST = ThreatExchangeExtensionManifest(
