+++ python-threatexchange/threatexchange/extensions/vpdq/vpdq_util.py
@@ -13,6 +13,8 @@ TIMESTAMP = "timestamp"
 VPDQ_TIMESTAMP_PRECISION = 3
 BITS_IN_VPDQ = 256
 VPDQ_VIDEOID_TYPE = t.Union[str, int]
+VPDQ_CONFIDENT_QUALITY_THRESHOLD = 50
+VPDQ_CONFIDENT_DISTANCE_THRESHOLD = 31
 
 
 @dataclass
